moved enums to seperate header file
