+++ c_source_files/chessboard.c
@@ -1,15 +1,7 @@
 #include <stdint.h>
 #include <stdbool.h>
 #include <stdlib.h>
-
-enum colored_chesspiece{w_King = 1, w_Queen, w_Bishop, w_Knight, w_Rook, w_Pawn, 
-                        b_King, b_Queen, b_Bishop, b_Knight, b_Rook, b_Pawn};
-
-enum chesspiece{no_Piece, King, Queen, Bishop, Knight, Rook, Pawn};
-
-enum color{white, black};
-
-enum event{no_Event, Castles, Promotion, En_passaint, Checkmate, Remis};
+#include "enums.h"
 
 const uint8_t init_chessboard[8][8] =
 {
